// prefect-ui/src/api/prefect.tsx
export interface components {
    schemas: {
// ...
        Flow: {
            /**
             * Id
             * Format: uuid
             */
            id?: string;
            /** Created */
            created?: string | null;
            /** Updated */
            updated?: string | null;
            /**
             * Name
             * @description The name of the flow
             */
            name: string;
            /**
             * Tags
             * @description A list of flow tags
             */
            tags?: string[];
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}